Define default theme to make a site health check warning go away
In WP admin, at URL /wp-admin/site-health.php, a `Have a default theme available`
message is annoying us all the time if (the default) WP_DEFAULT_THEME is missing
(which it is, since we've deleted it).

To hide the bogus warning, let's redefine the WP_DEFAULT_THEME from its default
value (`twenty...` something) to `pdkranj`.